pyproject.toml
[build-system]
requires = [
    "setuptools"
]
build-backend = "setuptools.build_meta"

[project]
name = "clipperbot"
version = "2.0"
description = "Discord bot that clips livestreams"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "discord.py @ git+https://github.com/Rapptz/discord.py.git@bdda31307b2892c0aafa2b8eb5bcfad7316452a9",
    "yt-dlp @ git+https://github.com/OdielDomanie/yt-dlp.git",
    "starlette",
    "uvicorn",
    "aiohttp",
    "python-dateutil",
    "psutil",
    "opencv-python-headless",
    "python-dotenv",
    "uvloop",
]

[tool.setuptools]
packages = ["clipperbot"]
